remove dependencies
corsica's 2.5 and up have a minimum node version that no longer needs the Promise polyfill. Fetch has been depreciated with an unclear successor but when I went to update the code I couldn't find where it was actually being used.

This change removes the polyfill, bumps the versions, and regenerates package-lock